XbrlTest - Portal improvements + currency converter in DevFunc
 - portal to drop and refresh problematic packages from cache (better)
 - move JsonValidation to a separate class
 - DevFunctions CurrencyConverter test and stats (looks good)

+++ dev/XbrlTest01/src/hu/sze/uni/xbrl/XbrlFilingManager.java
@@ -592,7 +592,7 @@ public class XbrlFilingManager implements XbrlConsts, DustUtilsConsts {
 							XbrlTestPortalUtils.extractWithApacheZipFile(fSrc, fRep, repName);
 						}
 					}
-					if ( null != fRep ) {
+					if ( (null != fRep ) && fRep.isFile()) {
 						try {
 							XbrlReportLoaderDomBase.createSplitCsv(fRep, dir, genFilePrefix, TEXT_CUT_AT);
 							fRep.delete();
